+++ Solution/SW.MB.Data/Contracts/Models/IPerson.cs
@@ -1,5 +1,5 @@
-﻿namespace SW.MB.Data.Contracts {
-  public interface IPerson : IEntity {
+﻿namespace SW.MB.Data.Contracts.Models {
+  public interface IPerson: IEntity {
     public string Firstname { get; }
     public string Lastname { get; }
     public DateTime? DateOfBirth { get; }
