+++ Solution/SW.MB.Data/Models/Entities/Abstracts/Entity.cs
@@ -1,7 +1,8 @@
-﻿using SW.MB.Data.Contracts;
+﻿using SW.MB.Data.Contracts.Models;
 
-namespace SW.MB.Data.Models.Entities.Abstracts {
-  public abstract class Entity: IEntity {
+namespace SW.MB.Data.Models.Entities.Abstracts
+{
+    public abstract class Entity: IEntity {
     public int ID { get; set; }
     public DateTime Created { get; set; }
     public int CreatedBy { get; set; }
