+++ Solution/SW.MB.Data/Models/Entities/Abstracts/PersonEntity.cs
@@ -1,7 +1,8 @@
-﻿using SW.MB.Data.Contracts;
+﻿using SW.MB.Data.Contracts.Models;
 
-namespace SW.MB.Data.Models.Entities.Abstracts {
-  public abstract class PersonEntity: Entity, IPerson {
+namespace SW.MB.Data.Models.Entities.Abstracts
+{
+    public abstract class PersonEntity: Entity, IPerson {
     public virtual ICollection<MandatorEntity>? Mandators { get; set; }
 
     public string Firstname { get; set; } = string.Empty;
