+++ Solution/SW.MB.Domain/Contracts/Services/ICompositionsService.cs
@@ -1,6 +1,6 @@
 ﻿using SW.MB.Domain.Models.Records;
 
-namespace SW.MB.Domain.Contracts {
+namespace SW.MB.Domain.Contracts.Services {
   public interface ICompositionsService {
     public IEnumerable<CompositionRecord> GetAll();
