+++ Solution/SW.MB.Domain/Contracts/Services/IMandatorsService.cs
@@ -1,6 +1,6 @@
 ﻿using SW.MB.Domain.Models.Records;
 
-namespace SW.MB.Domain.Contracts {
+namespace SW.MB.Domain.Contracts.Services {
   public interface IMandatorsService {
     public IEnumerable<MandatorRecord> GetAll();
   }
