+++ Solution/SW.MB.Domain/Contracts/Services/IMembersService.cs
@@ -1,6 +1,6 @@
 ﻿using SW.MB.Domain.Models.Records;
 
-namespace SW.MB.Domain.Contracts {
+namespace SW.MB.Domain.Contracts.Services {
   public interface IMembersService {
     public IEnumerable<MemberRecord> GetAll();
   }
