+++ Solution/SW.MB.Domain/Contracts/Services/IMusiciansService.cs
@@ -1,6 +1,6 @@
 ﻿using SW.MB.Domain.Models.Records;
 
-namespace SW.MB.Domain.Contracts {
+namespace SW.MB.Domain.Contracts.Services {
   public interface IMusiciansService {
     public IEnumerable<MusicianRecord> GetAll();
   }
