+++ Solution/SW.MB.Domain/Contracts/Services/IUsersService.cs
@@ -1,6 +1,6 @@
 ﻿using SW.MB.Domain.Models.Records;
 
-namespace SW.MB.Domain.Contracts {
+namespace SW.MB.Domain.Contracts.Services {
   public interface IUsersService {
     public IEnumerable<UserRecord> GetAll();
   }
