+++ Solution/SW.MB.Domain/Models/Records/Abstracts/EntityRecord.cs
@@ -1,7 +1,8 @@
-﻿using SW.MB.Data.Contracts;
+﻿using SW.MB.Data.Contracts.Models;
 
-namespace SW.MB.Domain.Models.Records.Abstracts {
-  public abstract record class EntityRecord : IEntity {
+namespace SW.MB.Domain.Models.Records.Abstracts
+{
+    public abstract record class EntityRecord : IEntity {
     public int ID { get; init; }
     public DateTime Created { get; init; } = DateTime.Now;
     public int CreatedBy { get; init; }
