+++ Solution/SW.MB.Domain/Models/Records/Abstracts/PersonRecord.cs
@@ -1,7 +1,8 @@
-﻿using SW.MB.Data.Contracts;
+﻿using SW.MB.Data.Contracts.Models;
 
-namespace SW.MB.Domain.Models.Records.Abstracts {
-  public abstract record class PersonRecord : EntityRecord, IPerson {
+namespace SW.MB.Domain.Models.Records.Abstracts
+{
+    public abstract record class PersonRecord : EntityRecord, IPerson {
     public string Firstname { get; init; } = string.Empty;
     public string Lastname { get; init; } = string.Empty;
     public DateTime? DateOfBirth { get; init; }
