+++ Solution/SW.MB.Domain/Services/DefaultCompositionsService.cs
@@ -1,12 +1,13 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using SW.MB.Data.Contracts;
-using SW.MB.Domain.Contracts;
+using SW.MB.Data.Contracts.UnitsOfWork;
+using SW.MB.Domain.Contracts.Services;
 using SW.MB.Domain.Extensions;
-using SW.MB.Domain.Implementations.Abstracts;
 using SW.MB.Domain.Models.Records;
+using SW.MB.Domain.Services.Abstracts;
 
-namespace SW.MB.Domain.Implementations {
-  internal class DefaultCompositionsService: ServiceBase, ICompositionsService {
+namespace SW.MB.Domain.Services
+{
+    internal class DefaultCompositionsService: ServiceBase, ICompositionsService {
     #region CONSTRUCTORS
     public DefaultCompositionsService(IServiceProvider serviceProvider) : base(serviceProvider) {
       // empty...
