+++ Solution/SW.MB.Domain/Services/DefaultMusiciansService.cs
@@ -1,12 +1,13 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using SW.MB.Data.Contracts;
-using SW.MB.Domain.Contracts;
+using SW.MB.Data.Contracts.UnitsOfWork;
+using SW.MB.Domain.Contracts.Services;
 using SW.MB.Domain.Extensions;
-using SW.MB.Domain.Implementations.Abstracts;
 using SW.MB.Domain.Models.Records;
+using SW.MB.Domain.Services.Abstracts;
 
-namespace SW.MB.Domain.Implementations {
-  internal class DefaultMusiciansService: ServiceBase, IMusiciansService {
+namespace SW.MB.Domain.Services
+{
+    internal class DefaultMusiciansService: ServiceBase, IMusiciansService {
     #region CONSTRUCTORS
     public DefaultMusiciansService(IServiceProvider serviceProvider) : base(serviceProvider) {
       // empty...
