+++ Solution/SW.MB.UI.WPF/ViewModels/AppViewModel.cs
@@ -4,12 +4,13 @@ using System.Linq;
 using System.Windows.Input;
 using Microsoft.Extensions.DependencyInjection;
 using SW.Framework.WPF;
-using SW.MB.Domain.Contracts;
+using SW.MB.Domain.Contracts.Services;
 using SW.MB.Domain.Models.Records;
 using SW.MB.UI.WPF.Models.Observables;
 
-namespace SW.MB.UI.WPF.ViewModels {
-  public class AppViewModel : ViewModel {
+namespace SW.MB.UI.WPF.ViewModels
+{
+    public class AppViewModel : ViewModel {
     public ObservableCollection<ObservableComposition> Compositions { get; } = new();
     public ObservableCollection<ObservableMandator> Mandators { get; } = new();
     public ObservableCollection<ObservableMember> Members { get; } = new();
